requirements.txt
flask
uk-covid19
newsapi-python
pytest
